D'hondt (missing votes analyzes)
A function or program designed to increase the total number of votes obtained by a political party in the Polish parliamentary elections of October 2023 by a user-specified amount, such as 100, 1000, or any designated figure, in every electoral district. Subsequently, the program utilizes a simulation of the D'Hondt vote distribution system to identify the district where the analyzed party came very close to securing an additional parliamentary mandate.

Through the use of this function, it becomes possible, for example, to determine that the party "Prawo i Sprawiedliwość" would have secured an extra-parliamentary mandate in District 33 (Świętokrzyskie) if the party's candidates had received just 102 more votes in that district.