Electoral commissions analysis
The provided Python code offers a combination of insights extracted during the analysis and the corresponding lines of code responsible for these findings. This analysis focused on a dataset retrieved from the PKW website, involving the processing and examination of electoral commissions from the parliamentary election in Poland, which took place in October 2023.

The analysis delved into various aspects, such as geographical breakdowns by voivodeships and communes, political party participation in vote distribution across different commission sizes,  and numerous other discoveries related to electoral commissions from the last election in Poland, which occurred on October 23.